Move e2e libs to separate path
Looks like having then under test breaks ginkgo output and
does not shows until the end

Also add ./test to vet and put correct package name at modules
