+++ test/reporter/reporter.go
@@ -1,4 +1,4 @@
-package e2e
+package reporter
 
 import (
 	"fmt"
