+++ test/runner/pod.go
@@ -1,4 +1,4 @@
-package e2e
+package runner
 
 import (
 	"strings"
